appboard/requirements.txt
Flask==3.1.0
requests==2.32.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyYAML==6.0.2
